feat(overlay): tasklog plugin
